Address a http://errorprone.info/bugpattern/NarrowingCompoundAssignment warning.
RELNOTES=n/a
PiperOrigin-RevId: 793778774

+++ guava-tests/test/com/google/common/cache/CacheTesting.java
@@ -170,9 +170,9 @@ final class CacheTesting {
     }
   }
 
-  static int getTotalSegmentSize(Cache<?, ?> cache) {
+  static long getTotalSegmentSize(Cache<?, ?> cache) {
     LocalCache<?, ?> map = toLocalCache(cache);
-    int totalSize = 0;
+    long totalSize = 0;
     for (Segment<?, ?> segment : map.segments) {
       totalSize += segment.maxSegmentWeight;
     }
